LayoutTests/inspector/indexeddb/deleteDatabaseNamesWithSpace.html

<!DOCTYPE html>
<html>
<head>
<script src="../../http/tests/inspector/resources/inspector-test.js"></script>
<script src="resources/utilities.js"></script>
<script>
function test()
{
    let suite = InspectorTest.createAsyncSuite("IndexedDB.requestDatabaseNames.spaces");

    suite.addTestCase({
        name: "ClearDatabases",
        description: "Remove any existing IndexedDB databases.",
        test: (resolve, reject) => {
            // FIXME: Remove any existing IndexedDB databases that might exist to workaround:
            // <https://webkit.org/b/148006> Each test should run with its own clean data store
            IndexedDBAgent.requestDatabaseNames(WebInspector.frameResourceManager.mainFrame.securityOrigin, (error, names) => {
                InspectorTest.evaluateInPage("deleteDatabaseNames(" + JSON.stringify(names) + ")", resolve);
            });
        }
    });

    suite.addTestCase({
        name: "EnsureNoDatabases",
        description: "Ensure no databases exist at the start.",
        test: (resolve, reject) => {
            // FIXME: <https://webkit.org/b/161285> Race between deleting a database and requesting database names seen in test
            setTimeout(() => {
                IndexedDBAgent.requestDatabaseNames(WebInspector.frameResourceManager.mainFrame.securityOrigin, (error, names) => {
                    InspectorTest.expectNoError(error);
                    InspectorTest.expectThat(names.length === 0, "No IndexedDB databases should exist initially.");
                    resolve();
                });
            }, 50);
        }
    });

    suite.addTestCase({
        name: "CreateDatabaseWithSpacesInName",
        description: "Create a database with spaces in the name.",
        test: (resolve, reject) => {
            InspectorTest.evaluateInPage("createEmptyDatabase('Database With Space')");
            IndexedDBAgent.requestDatabaseNames(WebInspector.frameResourceManager.mainFrame.securityOrigin, (error, names) => {
                InspectorTest.expectNoError(error);
                InspectorTest.expectThat(names.length === 1, "A single IndexedDB database should exist.");
                InspectorTest.log(JSON.stringify(names));
                resolve();
            });
        }
    });

    suite.addTestCase({
        name: "DeleteDatabaseWithSpacesInName",
        description: "Delete the database.",
        test: (resolve, reject) => {
            InspectorTest.evaluateInPage("deleteDatabaseNames(['Database With Space'])");
            // FIXME: <https://webkit.org/b/161285> Race between deleting a database and requesting database names seen in test
            setTimeout(() => {
                IndexedDBAgent.requestDatabaseNames(WebInspector.frameResourceManager.mainFrame.securityOrigin, (error, names) => {
                    InspectorTest.expectNoError(error);
                    InspectorTest.expectThat(names.length === 0, "No IndexedDB databases should exist because we just deleted them.");
                    InspectorTest.log(JSON.stringify(names));
                    resolve();
                });
            }, 50);
        }
    });

    suite.runTestCasesAndFinish();
}
</script>
</head>
<body onload="runTest()">
</body>
</html>
